+++ taf/developer_tool.py
@@ -151,21 +151,22 @@ def _load_signing_keys(
     return keys
 
 
-def _update_target_repos(repo_path, targets_dir, target_repo_path):
-    """Updates target repo's commit sha"""
+def _update_target_repos(repo_path, targets_dir, target_repo_path, add_branch):
+    """Updates target repo's commit sha and branch"""
     if not target_repo_path.is_dir() or target_repo_path == repo_path:
         return
-
     target_repo = GitRepository(str(target_repo_path))
     if target_repo.is_git_repository:
-        commit = target_repo.head_commit_sha()
+        data = {"commit": target_repo.head_commit_sha()}
+        if add_branch:
+            data["branch"] = target_repo.get_current_branch()
         target_repo_name = target_repo_path.name
-        (targets_dir / target_repo_name).write_text(
-            json.dumps({"commit": commit}, indent=4)
-        )
+        (targets_dir / target_repo_name).write_text(json.dumps(data, indent=4))
 
 
-def update_target_repos_from_fs(repo_path, targets_directory, namespace=None):
+def update_target_repos_from_fs(
+    repo_path, targets_directory, namespace=None, add_branch=True
+):
     """
     <Purpose>
         Create or update target files by reading the latest commits of the provided target repositories
@@ -185,27 +186,38 @@ def update_target_repos_from_fs(repo_path, targets_directory, namespace=None):
     if namespace:
         auth_repo_targets_dir = auth_repo_targets_dir / namespace
         auth_repo_targets_dir.mkdir(parents=True, exist_ok=True)
-
     for target_repo_path in targets_dir.glob("*"):
-        _update_target_repos(repo_path, auth_repo_targets_dir, target_repo_path)
+        _update_target_repos(
+            repo_path, auth_repo_targets_dir, target_repo_path, add_branch
+        )
 
 
-def update_target_repos_from_repositories_json(repo_path):
+def update_target_repos_from_repositories_json(
+    repo_path, targets_directory, add_branch=True
+):
     """
     <Purpose>
-        Create or update target files by reading the latest commits repositories.json
+        Create or update target files by reading the latest commit's repositories.json
     <Arguments>
         repo_path:
         Authentication repository's location
+        targets_directory:
+        Directory containing target repositories
     """
-    repo_path = Path(repo_path).resolve()
-    targets_dir = repo_path / TARGETS_DIRECTORY_NAME
-    repositories_json = json.loads(Path(targets_dir).read_text())
-
-    for repo in repositories_json.get("repositories"):
-        target_repo_path = targets_dir / repo
-        target_repo_path.mkdir(parents=True, exist_ok=True)
-        _update_target_repos(repo_path, targets_dir, target_repo_path)
+    auth_repo_path = Path(repo_path).resolve()
+    auth_repo_targets_dir = auth_repo_path / TARGETS_DIRECTORY_NAME
+    repositories_json = json.loads(
+        Path(auth_repo_targets_dir / "repositories.json").read_text()
+    )
+    targets_directory = Path(targets_directory).resolve()
+    for repo_name in repositories_json.get("repositories"):
+        target_repo_path = targets_directory / repo_name
+        namespace_and_name = repo_name.rsplit("/", 1)
+        if len(namespace_and_name) > 1:
+            namespace, _ = namespace_and_name
+            targets_dir = auth_repo_targets_dir / namespace
+        targets_dir.mkdir(parents=True, exist_ok=True)
+        _update_target_repos(repo_path, targets_dir, target_repo_path, add_branch)
 
 
 def build_auth_repo(
@@ -478,7 +490,6 @@ def generate_repositories_json(
 
     custom_data = read_input_dict(custom_data)
     repositories = {}
-
     repo_path = Path(repo_path).resolve()
     auth_repo_targets_dir = repo_path / TARGETS_DIRECTORY_NAME
     targets_directory = Path(targets_directory).resolve()
@@ -503,7 +514,7 @@ def generate_repositories_json(
         url = target_repo.get_remote_url()
         if url is None:
             if targets_relative_dir is not None:
-                url = Path(target_repo.repo_path).relative_to(targets_relative_dir)
+                url = Path(os.path.relpath(target_repo.repo_path, targets_relative_dir))
             else:
                 url = Path(target_repo.repo_path).resolve()
             # convert to posix path
@@ -686,7 +697,7 @@ def _write_targets_metadata(taf_repo, keystore, roles_key_infos, scheme):
 
     loaded_yubikeys = {}
     targets_keys = _load_signing_keys(
-        taf_repo, "targets", keystore, roles_key_infos, loaded_yubikeys
+        taf_repo, "targets", keystore, roles_key_infos, loaded_yubikeys, scheme=scheme
     )
     if len(targets_keys):
         taf_repo.update_targets_from_keystore(targets_keys[0], write=False)
@@ -696,9 +707,11 @@ def _write_targets_metadata(taf_repo, keystore, roles_key_infos, scheme):
                 pin = yk.get_key_pin(serial_num)
                 taf_repo.update_targets(pin, write=False)
                 break
-    snapshot_keys = _load_signing_keys(taf_repo, "snapshot", keystore, roles_key_infos)
+    snapshot_keys = _load_signing_keys(
+        taf_repo, "snapshot", keystore, roles_key_infos, scheme=scheme
+    )
     timestamp_keys = _load_signing_keys(
-        taf_repo, "timestamp", keystore, roles_key_infos
+        taf_repo, "timestamp", keystore, roles_key_infos, scheme=scheme
     )
     snapshot_key = snapshot_keys[0]
     timestamp_key = timestamp_keys[0]